refactor: 네트워크 통신 관련 코드 수정
