+++ app/src/main/java/woowacourse/shopping/data/mapper/CartProductMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.mapper
 
 import woowacourse.shopping.data.cart.CartProductEntity
-import woowacourse.shopping.domain.CartProduct
+import woowacourse.shopping.domain.model.CartProduct
 
 fun CartProductEntity.toDomain() =
     CartProduct(
