+++ app/src/main/java/woowacourse/shopping/data/mapper/RecentProductMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.mapper
 
 import woowacourse.shopping.data.recent.RecentProductEntity
-import woowacourse.shopping.domain.RecentProduct
+import woowacourse.shopping.domain.model.RecentProduct
 
 fun RecentProductEntity.toDomain() =
     RecentProduct(
