+++ app/src/main/java/woowacourse/shopping/data/product/ProductRepositoryImpl.kt
@@ -2,12 +2,18 @@ package woowacourse.shopping.data.product
 
 import woowacourse.shopping.data.PagedResult
 import woowacourse.shopping.data.product.remote.ProductRemoteDataSource
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.repository.ProductRepository
+import kotlin.concurrent.thread
 
 class ProductRepositoryImpl(
     private val remoteDataSource: ProductRemoteDataSource,
 ) : ProductRepository {
-    override fun getAll(): List<Product> = remoteDataSource.getAll()
+    override fun getAll(): List<Product> {
+        var result = emptyList<Product>()
+        thread { result = remoteDataSource.getAll() }.join()
+        return result
+    }
 
     override fun getProductById(id: Long): Product? = remoteDataSource.getProductById(id)
 
@@ -17,6 +23,8 @@ class ProductRepositoryImpl(
     ): PagedResult<Product> {
         require(offset >= 0)
         require(limit > 0)
-        return remoteDataSource.getPagedProducts(limit, offset)
+        var result = PagedResult<Product>(emptyList(), false)
+        thread { result = remoteDataSource.getPagedProducts(limit, offset) }.join()
+        return result
     }
 }
