+++ app/src/main/java/woowacourse/shopping/domain/model/CartProduct.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 data class CartProduct(
     val id: Long? = null,
