+++ app/src/main/java/woowacourse/shopping/domain/model/Product.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 import java.io.Serializable
