+++ app/src/main/java/woowacourse/shopping/domain/model/RecentProduct.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.model
 
 data class RecentProduct(
     val id: Long? = null,
