+++ app/src/main/java/woowacourse/shopping/domain/repository/ProductRepository.kt
@@ -1,7 +1,7 @@
-package woowacourse.shopping.data.product
+package woowacourse.shopping.domain.repository
 
 import woowacourse.shopping.data.PagedResult
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.Product
 
 interface ProductRepository {
     fun getAll(): List<Product>
