+++ app/src/main/java/woowacourse/shopping/view/cart/ShoppingCartViewModel.kt
@@ -3,10 +3,10 @@ package woowacourse.shopping.view.cart
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.cart.CartProductRepository
-import woowacourse.shopping.domain.CartProduct
-import woowacourse.shopping.domain.Product
-import woowacourse.shopping.domain.ShoppingCart
+import woowacourse.shopping.domain.model.CartProduct
+import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.model.ShoppingCart
+import woowacourse.shopping.domain.repository.CartProductRepository
 
 class ShoppingCartViewModel(
     private val repository: CartProductRepository,
