+++ app/src/main/java/woowacourse/shopping/view/common/QuantityControlEventHandler.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.view.common
 
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.Product
 
 interface QuantityControlEventHandler {
     fun onQuantityIncreaseClick(item: Product)
