+++ app/src/main/java/woowacourse/shopping/view/product/catalog/adapter/RecentProductViewHolder.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemRecentProductBinding
-import woowacourse.shopping.domain.RecentProduct
+import woowacourse.shopping.domain.model.RecentProduct
 
 class RecentProductViewHolder(
     private val binding: ItemRecentProductBinding,
