+++ app/src/main/java/woowacourse/shopping/view/product/detail/ProductDetailActivity.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.ViewModelProvider
 import woowacourse.shopping.R
 import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.databinding.ActivityProductDetailBinding
-import woowacourse.shopping.domain.Product
+import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.view.cart.ShoppingCartActivity
 import woowacourse.shopping.view.util.getSerializableExtraCompat
