+++ app/src/main/res/layout/activity_product_detail.xml
@@ -13,11 +13,11 @@
 
         <variable
             name="product"
-            type="woowacourse.shopping.domain.Product" />
+            type="woowacourse.shopping.domain.model.Product" />
 
         <variable
             name="lastProduct"
-            type="woowacourse.shopping.domain.RecentProduct" />
+            type="woowacourse.shopping.domain.model.RecentProduct" />
 
         <variable
             name="handler"
