+++ app/src/main/res/layout/item_product.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="product"
-            type="woowacourse.shopping.domain.Product" />
+            type="woowacourse.shopping.domain.model.Product" />
 
         <variable
             name="quantity"
