+++ app/src/main/res/layout/item_quantity_control.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="product"
-            type="woowacourse.shopping.domain.Product" />
+            type="woowacourse.shopping.domain.model.Product" />
 
         <variable
             name="quantity"
