+++ app/src/main/res/layout/item_recent_product.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="recentProduct"
-            type="woowacourse.shopping.domain.RecentProduct" />
+            type="woowacourse.shopping.domain.model.RecentProduct" />
 
         <variable
             name="handler"
